requirements.txt
botbuilder-integration-aiohttp>=4.11.0
botbuilder-dialogs>=4.11.0
botbuilder-ai>=4.11.0
datatypes-date-time>=1.0.0.a2
pyodbc>=4.0.30
pyadaptivecards>=0.1.1
azure-ai-textanalytics
botbuilder-schema>=4.11.0
botframework-connector>=4.11.0
azure-mgmt-resource
azure-mgmt-storage
azure-identity
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0